test: move ESM fixtures to fixtures dir
Also consistently import the `common` module where possible.
