--- test/es-module/esm-snapshot.js
@@ -1,3 +1,0 @@
-/* eslint-disable node-core/required-modules */
-'use strict';
-module.exports = 1;
