+++ test/es-module/test-esm-basic-imports.mjs
@@ -1,7 +1,7 @@
 // Flags: --experimental-modules
 import '../common';
 import assert from 'assert';
-import ok from './test-esm-ok.mjs';
+import ok from '../fixtures/es-modules/test-esm-ok.mjs';
 import okShebang from './test-esm-shebang.mjs';
 
 assert(ok);
