+++ test/es-module/test-esm-double-encoding.mjs
@@ -3,4 +3,4 @@ import '../common';
 
 // Assert we can import files with `%` in their pathname.
 
-import './test-esm-double-encoding-native%2520.js';
+import '../fixtures/es-modules/test-esm-double-encoding-native%2520.js';
