+++ test/es-module/test-esm-dynamic-import.js
@@ -7,8 +7,8 @@ const vm = require('vm');
 
 common.crashOnUnhandledRejection();
 
-const relativePath = './test-esm-ok.mjs';
-const absolutePath = require.resolve('./test-esm-ok.mjs');
+const relativePath = '../fixtures/es-modules/test-esm-ok.mjs';
+const absolutePath = require.resolve('../fixtures/es-modules/test-esm-ok.mjs');
 const targetURL = new URL('file:///');
 targetURL.pathname = absolutePath;
