+++ test/es-module/test-esm-encoded-path.mjs
@@ -2,6 +2,6 @@
 import '../common';
 import assert from 'assert';
 // ./test-esm-ok.mjs
-import ok from './test-%65%73%6d-ok.mjs';
+import ok from '../fixtures/es-modules/test-%65%73%6d-ok.mjs';
 
 assert(ok);
