+++ test/es-module/test-esm-example-loader.js
@@ -1,6 +1,6 @@
 // Flags: --experimental-modules --loader ./test/fixtures/es-module-loaders/example-loader.mjs
 /* eslint-disable node-core/required-modules */
 import assert from 'assert';
-import ok from './test-esm-ok.mjs';
+import ok from '../fixtures/es-modules/test-esm-ok.mjs';
 
 assert(ok);
