+++ test/es-module/test-esm-loader-dependency.mjs
@@ -1,5 +1,5 @@
 // Flags: --experimental-modules --loader ./test/fixtures/es-module-loaders/loader-with-dep.mjs
 /* eslint-disable node-core/required-modules */
-import './test-esm-ok.mjs';
+import '../fixtures/es-modules/test-esm-ok.mjs';
 
 // We just test that this module doesn't fail loading
