+++ test/es-module/test-esm-main-lookup.mjs
@@ -1,5 +1,5 @@
 // Flags: --experimental-modules
-/* eslint-disable node-core/required-modules */
+import '../common';
 import assert from 'assert';
 import main from '../fixtures/es-modules/pjson-main';
