+++ test/es-module/test-esm-named-exports.mjs
@@ -1,9 +1,8 @@
 // Flags: --experimental-modules --loader ./test/fixtures/es-module-loaders/builtin-named-exports-loader.mjs
-/* eslint-disable node-core/required-modules */
-import '../common/index';
+import '../common';
 import { readFile } from 'fs';
 import assert from 'assert';
-import ok from './test-esm-ok.mjs';
+import ok from '../fixtures/es-modules/test-esm-ok.mjs';
 
 assert(ok);
 assert(readFile);
