+++ test/es-module/test-esm-namespace.mjs
@@ -1,7 +1,5 @@
 // Flags: --experimental-modules
-/* eslint-disable node-core/required-modules */
-
-import '../common/index';
+import '../common';
 import * as fs from 'fs';
 import assert from 'assert';
