--- test/es-module/test-esm-ok.mjs
@@ -1,5 +1,0 @@
-// Flags: --experimental-modules
-/* eslint-disable node-core/required-modules */
-
-const isJs = true;
-export default isJs;
