+++ test/es-module/test-esm-preserve-symlinks.js
@@ -1,4 +1,3 @@
-// Flags: --experimental-modules
 'use strict';
 
 const common = require('../common');
