+++ test/es-module/test-esm-resolve-hook.mjs
@@ -2,7 +2,7 @@
 /* eslint-disable node-core/required-modules */
 import { namedExport } from '../fixtures/es-module-loaders/js-as-esm.js';
 import assert from 'assert';
-import ok from './test-esm-ok.mjs';
+import ok from '../fixtures/es-modules/test-esm-ok.mjs';
 
 assert(ok);
 assert(namedExport);
