+++ test/es-module/test-esm-shared-loader-dep.mjs
@@ -1,7 +1,7 @@
 // Flags: --experimental-modules --loader ./test/fixtures/es-module-loaders/loader-shared-dep.mjs
-/* eslint-disable node-core/required-modules */
+import '../common';
 import assert from 'assert';
-import './test-esm-ok.mjs';
+import '../fixtures/es-modules/test-esm-ok.mjs';
 import dep from '../fixtures/es-module-loaders/loader-dep.js';
 
 assert.strictEqual(dep.format, 'esm');
