+++ test/es-module/test-esm-shebang.mjs
@@ -1,6 +1,6 @@
 #! }]) // isn't js
 // Flags: --experimental-modules
-/* eslint-disable node-core/required-modules */
+import '../common';
 
 const isJs = true;
 export default isJs;
