+++ test/fixtures/es-modules/esm-snapshot-mutator.js
@@ -1,4 +1,3 @@
-/* eslint-disable node-core/required-modules */
 'use strict';
 const shouldSnapshotFilePath = require.resolve('./esm-snapshot.js');
 require('./esm-snapshot.js');
