+++ test/fixtures/es-modules/esm-snapshot.js
@@ -1,0 +1,2 @@
+'use strict';
+module.exports = 1;
