+++ test/fixtures/es-modules/json.json
