+++ test/fixtures/es-modules/test-esm-double-encoding-native%20.js
@@ -1,5 +1,4 @@
 'use strict';
-require('../common');
 
 // Trivial test to assert we can load files with `%` in their pathname.
 // Imported by `test-esm-double-encoding.mjs`.
