+++ test/fixtures/es-modules/test-esm-ok.mjs
@@ -1,0 +1,2 @@
+const isJs = true;
+export default isJs;
