+++ test/parallel/test-module-main-extension-lookup.js
@@ -1,9 +1,11 @@
 'use strict';
 require('../common');
+const fixtures = require('../common/fixtures');
 const { execFileSync } = require('child_process');
 
 const node = process.argv[0];
 
-execFileSync(node, ['--experimental-modules', 'test/es-module/test-esm-ok']);
 execFileSync(node, ['--experimental-modules',
-                    'test/fixtures/es-modules/noext']);
+                    fixtures.path('es-modules', 'test-esm-ok')]);
+execFileSync(node, ['--experimental-modules',
+                    fixtures.path('es-modules', 'noext')]);
